+++ src/test/kotlin/tornadofx/tests/TableViewTest.kt
@@ -1,20 +1,17 @@
 package tornadofx.tests
 
+import javafx.beans.property.ObjectProperty
 import javafx.beans.property.SimpleDoubleProperty
 import javafx.beans.property.SimpleIntegerProperty
 import javafx.beans.property.SimpleStringProperty
 import javafx.collections.FXCollections
 import javafx.scene.Scene
-import javafx.scene.input.MouseButton
 import javafx.scene.layout.StackPane
 import javafx.stage.Stage
-import org.junit.Before
 import org.junit.Test
 import org.testfx.api.FxRobot
 import org.testfx.api.FxToolkit
-import tornadofx.column
-import tornadofx.makeIndexColumn
-import tornadofx.tableview
+import tornadofx.*
 import java.nio.file.Paths
 
 class TableViewTest {
@@ -35,7 +32,9 @@ class TableViewTest {
                 tableview(TestList) {
                     makeIndexColumn()
                     column("A Column", TestObject::A)
-                    column("B Column", TestObject::B)
+                    column("B Column", Double::class) {
+                        value { it.value.B }
+                    }
                     column("C Column", TestObject::C)
                 }
                 setPrefSize(400.0, 160.0)